Added these files via upload
These are all the files to run our CSUBatch scheduler and dispatcher designed in C along with the jobs. The project is incomplete, we didn't get it fully working but we did the majority of it. The focus was on the software engineering process more than the technical side so we still received a good grade.